custom-elements/instantiating-custom-elements/changing-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Changing IS attribute of the custom element after element is instatiated</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="After a custom element is instantiated, changing the value of the is attribute must not affect this element's custom element type">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(HTML5_ELEMENTS[i]) == -1) {
            var element = HTML5_ELEMENTS[i];
            var doc = newHTMLDocument();
            
            var name = 'x-' + element;
            
            doc.register(name);
            
            if (HTML5_TABLE_ELEMENTS.indexOf(element) != -1) {
                doc.body.innerHTML = 
                    '<table>' +
                    '<' + element + ' id="customElement" is="' + name + '"></' + element + '>' +
                    '</table>';
            } else {
                doc.body.innerHTML = 
                    '<' + element + ' id="customElement" is="' + name + '"></' + element + '>';        
            }
            
            var customElement = doc.querySelector('#customElement');

            var obj = doc.createElement(element);
            
            assert_class_string(customElement, obj.constructor.name,
                    'Custom element  must be a ' + obj.constructor.name);
            
            for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
                var name2 = 'x-' + element + '-' + HTML5_ELEMENTS[i];        
                doc.register(name2);

                customElement.setAttribute('is', name2);
                
                assert_class_string(customElement, obj.constructor.name,
                        'Custom element must not be changed and should be a ' + 
                        obj.constructor.name);
            }
            
        }
    }
}, 'Changing IS attribute of the custom element after element is instatiated');

</script>
</body>
</html>
